+++ loganom/report_mail.py
@@ -1,6 +1,10 @@
 """Reports by email
 """
 
+# Workaround Python3.8 issue
+# https://github.com/marrow/mailer/issues/87
+import sys
+sys.modules["cgi.parse_qsl"] = None
 from marrow.mailer import Message, Mailer